remote-render/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "small-desktop-display-remote-render"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "pillow>=10.0.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
